SIH-25/ayushcare/backend/requirements.txt
Django==5.2          # LTS (recommended) — stable and supported. :contentReference[oaicite:5]{index=5}
djangorestframework==3.15
djangorestframework-simplejwt==2.8.0
channels==4.3.1      # Channels for WebSockets - works with Django 4.2+ / Python 3.9+. :contentReference[oaicite:6]{index=6}
channels-redis==4.0.0
celery==5.5.0        # Celery 5.x works with Python 3.8+; 5.5+ recommended. :contentReference[oaicite:7]{index=7}
redis==5.0.10
mysqlclient==2.1.1
pyfcm==1.2.8
django-cors-headers==4.0.0
python-dotenv==1.0.0
